# docs/openapi.yaml
openapi: 3.1.0
info:
  title: THCPN IoT Research Platform API
  version: 0.15.0
  description: |
    THCPN 科研物联网数据管理平台后端 API。

    当前文档覆盖已实现的阶段功能：健康检查、短信验证码登录/注册、邮箱验证码验证、密码注册/登录、TOTP MFA、refresh token、logout、会话管理、
    JWT 认证、Workspace 列表/创建、Workspace 成员管理、Project、Site、Device、DeviceRelation、DataStream、DataSource、DataStreamBinding、Telemetry Query、Media Query、Dataset 元信息与遥测查询管理和 ExportJob 导出任务、
    AccessGrant 外部分享/售后临时授权、Invitation、AuditLog、Camera 实时预览，以及系统管理员后台 DataSource / THCPN 标准站和组网站拓扑同步接口。

    正式业务接口优先使用 `Authorization: Bearer <token>`。
    开发环境可通过配置启用 `X-User-ID` fallback；生产环境应关闭该能力。
servers:
  - url: http://127.0.0.1:8080
    description: Local development API server
...
paths:
...
  /api/v1/auth/sms/login:
    post:
      tags:
        - Auth
      operationId: loginWithSms
      summary: Login or register with SMS verification code
      description: |
        Verifies the SMS code and issues a JWT access token plus a refresh token.
        If the phone number does not belong to an active user, a new user, personal workspace and Owner membership are created.
        Existing users with TOTP MFA enabled must also supply `mfa_code`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmsLoginRequest"
            examples:
              default:
                value:
                  phone: "13800000000"
                  code: "123456"
                  mfa_code: "654321"
                  name: 张三
      responses:
        "200":
          description: SMS login succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
              examples:
                loggedIn:
                  value:
                    access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    refresh_token: wJalrXUtnFEMI-K7MDENG_bPxRfiCYEXAMPLE
                    token_type: Bearer
                    expires_in: 86400
                    refresh_expires_in: 2592000
                    user:
                      id: 018fa0ec-9b6e-7a3d-9c7b-1d8b04f31988
                      name: 张三
                      phone: "13800000000"
                      status: active
                      phone_verified_at: "2026-06-30T14:12:54+08:00"
                      last_login_at: "2026-06-30T14:12:54+08:00"
                    created: true
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/auth/password/register:
    post:
      tags:
        - Auth
      operationId: registerWithPassword
      summary: Register with password
      description: |
        Creates a user, personal workspace, Owner membership and password credential in one transaction,
        then issues a JWT access token plus a refresh token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordRegisterRequest"
            examples:
              default:
                value:
                  name: 张三
                  phone: "13800000000"
                  email: zhangsan@example.com
                  password: Str0ngPassword!
      responses:
        "201":
          description: Password registration succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
              examples:
                registered:
                  value:
                    access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    refresh_token: wJalrXUtnFEMI-K7MDENG_bPxRfiCYEXAMPLE
                    token_type: Bearer
                    expires_in: 86400
                    refresh_expires_in: 2592000
                    user:
                      id: 018fa0ec-9b6e-7a3d-9c7b-1d8b04f31988
                      name: 张三
                      phone: "13800000000"
                      email: zhangsan@example.com
                      status: active
                      last_login_at: "2026-06-30T14:12:44+08:00"
                    created: true
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/auth/password/login:
    post:
      tags:
        - Auth
      operationId: loginWithPassword
      summary: Login with phone/email and password
      description: |
        Authenticates an existing active user by phone or email and password, then issues a JWT access token plus a refresh token.
        Repeated password failures increment the credential failure count and may temporarily lock the account.
        Accounts with TOTP MFA enabled must also supply `mfa_code`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordLoginRequest"
            examples:
              phone:
                value:
                  identifier: "13800000000"
                  password: Str0ngPassword!
                  mfa_code: "654321"
              email:
                value:
                  identifier: zhangsan@example.com
                  password: Str0ngPassword!
                  mfa_code: "654321"
      responses:
        "200":
          description: Password login succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
              examples:
                loggedIn:
                  value:
                    access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    refresh_token: wJalrXUtnFEMI-K7MDENG_bPxRfiCYEXAMPLE
                    token_type: Bearer
                    expires_in: 86400
                    refresh_expires_in: 2592000
                    user:
                      id: 018fa0ec-9b6e-7a3d-9c7b-1d8b04f31988
                      name: 张三
                      phone: "13800000000"
                      email: zhangsan@example.com
                      status: active
                      last_login_at: "2026-06-30T14:12:44+08:00"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/auth/password/complete-initial:
    post:
      tags:
        - Auth
      operationId: completeInitialPassword
      summary: Complete the first password change
      description: Consumes the one-time token returned after a temporary-password login and issues normal access and refresh tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteInitialPasswordRequest"
      responses:
        "200":
          description: Password changed and normal session issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/Internal"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token returned by SMS or password login APIs.
    devUserHeader:
      type: apiKey
      in: header
      name: X-User-ID
      description: Development-only fallback authentication. Production should disable it.
...
  responses:
    InvalidArgument:
      description: Invalid request argument or body.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            invalid:
              value:
                error:
                  code: INVALID_ARGUMENT
                  message: invalid request body
                  request_id: req_123
    Unauthorized:
      description: Missing, invalid or inactive authentication.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            unauthorized:
              value:
                error:
                  code: UNAUTHORIZED
                  message: invalid bearer token
                  request_id: req_123
...
    Conflict:
      description: Request conflicts with existing data or invariants.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            conflict:
              value:
                error:
                  code: CONFLICT
                  message: phone or email already exists
                  request_id: req_123
    RateLimited:
      description: Request was rate-limited or the account is temporarily locked.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            limited:
              value:
                error:
                  code: RATE_LIMITED
                  message: sms code requested too frequently
                  request_id: req_123
...
    Internal:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            internal:
              value:
                error:
                  code: INTERNAL
                  message: internal error
                  request_id: req_123
  schemas:
...
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
    ErrorBody:
      type: object
      required:
        - code
        - message
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
        request_id:
          type: string
    ErrorCode:
      type: string
      enum:
        - INVALID_ARGUMENT
        - UNAUTHORIZED
        - PERMISSION_DENIED
        - NOT_FOUND
        - CONFLICT
        - RATE_LIMITED
        - DATA_SOURCE_ERROR
        - INTERNAL
        - SERVICE_UNAVAILABLE
    UUID:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
    Phone:
      type: string
      description: Normalized phone number. Lightweight validation allows digits and an optional leading plus sign.
      examples:
        - "13800000000"
        - "+8613800000000"
    Email:
      type: string
      format: email
    UserStatus:
      type: string
      enum:
        - active
        - disabled
...
    SmsLoginRequest:
      type: object
      required:
        - phone
        - code
      properties:
        phone:
          $ref: "#/components/schemas/Phone"
        code:
          type: string
          pattern: ^[0-9]{6}$
          example: "123456"
        mfa_code:
          type: string
          pattern: ^[0-9]{6}$
          description: Required when the existing user has TOTP MFA enabled.
          example: "654321"
        name:
          type: string
          description: Used as the display name when SMS login auto-registers a new user.
          example: 张三
    PasswordRegisterRequest:
      type: object
      required:
        - password
      anyOf:
        - required:
            - phone
        - required:
            - email
      properties:
        name:
          type: string
          example: 张三
        phone:
          $ref: "#/components/schemas/Phone"
        email:
          $ref: "#/components/schemas/Email"
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          description: Must contain at least one letter and one digit.
          example: Str0ngPassword!
    PasswordLoginRequest:
      type: object
      required:
        - identifier
        - password
      properties:
        identifier:
          type: string
          description: Phone number or email address.
          examples:
            - "13800000000"
            - zhangsan@example.com
        password:
          type: string
          format: password
          example: Str0ngPassword!
        mfa_code:
          type: string
          pattern: ^[0-9]{6}$
          description: Required when TOTP MFA is enabled for the account.
          example: "654321"
...
    CompleteInitialPasswordRequest:
      type: object
      required:
        - password_change_token
        - password
      properties:
        password_change_token:
          type: string
          description: One-time short-lived token returned after temporary-password login.
        password:
          type: string
          format: password
...
    LoginResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - token_type
        - expires_in
        - refresh_expires_in
        - user
      properties:
        access_token:
          type: string
          description: JWT access token.
        refresh_token:
          type: string
          description: Opaque refresh token. Store securely and send only to `/api/v1/auth/refresh` or `/api/v1/auth/logout`.
        token_type:
          type: string
          enum:
            - Bearer
        expires_in:
          type: integer
          format: int64
          description: Access token lifetime in seconds.
        refresh_expires_in:
          type: integer
          format: int64
          description: Refresh token lifetime in seconds.
        user:
          $ref: "#/components/schemas/UserProfile"
        created:
          type: boolean
          description: True when the request created a new user.
        password_change_required:
          type: boolean
          description: True when the login only returned a short-lived password-change token.
        password_change_token:
          type: string
          description: One-time token used by complete-initial password endpoint.
        password_change_expires_in:
          type: integer
          format: int64
...
    UserProfile:
      type: object
      required:
        - id
        - name
        - status
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
        phone:
          $ref: "#/components/schemas/Phone"
        email:
          $ref: "#/components/schemas/Email"
        status:
          $ref: "#/components/schemas/UserStatus"
        phone_verified_at:
          $ref: "#/components/schemas/Timestamp"
        email_verified_at:
          $ref: "#/components/schemas/Timestamp"
        last_login_at:
          $ref: "#/components/schemas/Timestamp"
...